backend/requirements.txt
langchain==0.3.27
langchain-groq==0.2.1
langchain-community==0.3.27
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pygithub==2.5.0
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
openapi-generator-cli==4.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.4
httpx==0.28.1
aiofiles==24.1.0
zipfile-deflate64==0.2.0
python-dotenv==1.0.1
pydantic==2.10.3
requests==2.32.3
jinja2==3.1.4
markdown==3.7
